requirements.txt
pyyaml>=6.0
feedparser>=6.0
tweepy>=4.14.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
